created shiny app, edited which pairs should i breed
new shiny app which will be default modality through which this software is used.

which pairs should I breed's text output was slightly edited to be more precise. In the breeding scheme, even if the user wants a het in the end at a locus, it would be ideal to produce a homopos in the step before so as to maximize chances of obtaining a het in the next cross. Language now reflects this.